+++ license.txt
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) 2017 Vjacheslav Trushkin
+Copyright (c) 2016 - 2018 Vjacheslav Trushkin
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
